refactor(orders,freight): create server entrypoints, reduce allowlist by 7
Server entrypoints created/extended:
- src/modules/pedidos/server.ts — added:
  getRecentOrdersForCustomer, getOrderAggregate,
  findOrderWithShipmentByPrefix (from order.repository)
- src/modules/freight/server.ts — new file, exports:
  getQuoteContext (quote-context.repository),
  getShipmentsForOrder, getShipmentById (shipment-linkage.repository),
  createShipmentFromQuote, CreateShipmentInput (adapters/melhor-envio-shipment)

Consumers redirected (8 files):
- 4 frank tools → @/modules/pedidos/server + @/modules/freight/server
  (was pedidos/order.repository + freight/*.repository)
- 1 atendimento → @/modules/pedidos/server
  (was pedidos/order.repository)
- 2 freight routes → @/modules/freight/server
  (was freight/adapters/melhor-envio-shipment)

Allowlist reduced: 15 → 8 entries (removed 7):
- 4 frank tools
- 1 atendimento orchestrator
- 2 freight shipment routes

Remaining 8 entries (tracked debt):
- 2 test files (vi.mock patterns, not runtime imports)
- 6 UI cross-module (pending separate PR)

Zero functional changes.

QG: all 12 passed, RELEASE_CHECK_PASS
