+++ src/modules/frank/tools/read-only/getRecentOrders.tool.ts
@@ -1,5 +1,5 @@
-import { getRecentOrdersForCustomer } from '@/modules/pedidos/order.repository';
-import { getShipmentsForOrder } from '@/modules/freight/shipment-linkage.repository';
+import { getRecentOrdersForCustomer } from '@/modules/pedidos/server';
+import { getShipmentsForOrder } from '@/modules/freight/server';
 import { executeFrankTool } from '../tool-guard';
 import { clampSupportLimit, selectLatestShipment, toShipmentSummary, type ShipmentSummary } from './shared';
